# Optimization of data loading on auth reset