--- a/src/domain/collaboration/callout-contribution/callout.contribution.service.authorization.ts
+++ b/src/domain/collaboration/callout-contribution/callout.contribution.service.authorization.ts
@@ -32,27 +32,61 @@ export class CalloutContributionAuthorizationService {
   ) {}
 
   public async applyAuthorizationPolicy(
-    contributionInput: ICalloutContribution,
+    contributionID: string,
     parentAuthorization: IAuthorizationPolicy | undefined,
     communityPolicy: ICommunityPolicy,
     spaceSettings: ISpaceSettings
   ): Promise<IAuthorizationPolicy[]> {
     const contribution =
       await this.contributionService.getCalloutContributionOrFail(
-        contributionInput.id,
+        contributionID,
         {
+          loadEagerRelations: false,
           relations: {
+            authorization: true,
             post: {
+              authorization: true,
               profile: true,
               comments: {
                 authorization: true,
               },
             },
             whiteboard: {
+              authorization: true,
               profile: true,
             },
             link: {
+              authorization: true,
               profile: true,
+            },
+          },
+          select: {
+            id: true,
+            authorization:
+              this.authorizationPolicyService.authorizationSelectOptions,
+            whiteboard: {
+              id: true,
+              profile: {
+                id: true,
+              },
+              authorization:
+                this.authorizationPolicyService.authorizationSelectOptions,
+            },
+            post: {
+              id: true,
+              profile: {
+                id: true,
+              },
+              authorization:
+                this.authorizationPolicyService.authorizationSelectOptions,
+            },
+            link: {
+              id: true,
+              profile: {
+                id: true,
+              },
+              authorization:
+                this.authorizationPolicyService.authorizationSelectOptions,
             },
           },
         }