--- a/src/domain/collaboration/callout-contribution/callout.contribution.service.authorization.ts
+++ b/src/domain/collaboration/callout-contribution/callout.contribution.service.authorization.ts
@@ -68,10 +68,12 @@ export class CalloutContributionAuthorizationService {
           },
           select: {
             id: true,
+            createdBy: true,
             authorization:
               this.authorizationPolicyService.authorizationSelectOptions,
             post: {
               id: true,
+              createdBy: true,
               authorization:
                 this.authorizationPolicyService.authorizationSelectOptions,
               profile: {
@@ -87,6 +89,7 @@ export class CalloutContributionAuthorizationService {
             },
             whiteboard: {
               id: true,
+              createdBy: true,
               authorization:
                 this.authorizationPolicyService.authorizationSelectOptions,
               profile: {