--- a/src/domain/collaboration/callout-framing/callout.framing.service.authorization.ts
+++ b/src/domain/collaboration/callout-framing/callout.framing.service.authorization.ts
@@ -51,7 +51,7 @@ export class CalloutFramingAuthorizationService {
 
     const framingAuthorizations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        calloutFraming.profile,
+        calloutFraming.profile.id,
         calloutFraming.authorization
       );
     updatedAuthorizations.push(...framingAuthorizations);