--- a/src/domain/collaboration/innovation-flow/innovation.flow.service.authorization.ts
+++ b/src/domain/collaboration/innovation-flow/innovation.flow.service.authorization.ts
@@ -48,7 +48,7 @@ export class InnovationFlowAuthorizationService {
 
     const profileAuthorizations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        innovationFlow.profile,
+        innovationFlow.profile.id,
         innovationFlow.authorization
       );
     updatedAuthorizations.push(...profileAuthorizations);