--- a/src/domain/collaboration/post/post.service.authorization.ts
+++ b/src/domain/collaboration/post/post.service.authorization.ts
@@ -80,7 +80,7 @@ export class PostAuthorizationService {
     // cascade
     const profileAuthorizations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        post.profile,
+        post.profile.id,
         post.authorization
       );
     updatedAuthorizations.push(...profileAuthorizations);