--- a/src/domain/common/authorization-policy/authorization.policy.service.ts
+++ b/src/domain/common/authorization-policy/authorization.policy.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { FindOptionsSelect, Repository } from 'typeorm';
 import {
   EntityNotFoundException,
   ForbiddenException,
@@ -36,7 +36,7 @@ export class AuthorizationPolicyService {
     private readonly logger: LoggerService
   ) {}
 
-  public authorizationSelectOptions = {
+  public authorizationSelectOptions: FindOptionsSelect<AuthorizationPolicy> = {
     id: true,
     anonymousReadAccess: true,
     credentialRules: true,