--- a/src/domain/common/profile/profile.service.authorization.ts
+++ b/src/domain/common/profile/profile.service.authorization.ts
@@ -24,27 +24,21 @@ export class ProfileAuthorizationService {
       loadEagerRelations: false,
       relations: {
         authorization: true,
-        tagsets: {
-          authorization: true,
-        },
-        references: {
-          authorization: true,
-        },
-        visuals: {
-          authorization: true,
-        },
+        tagsets: { authorization: true },
+        references: { authorization: true },
+        visuals: { authorization: true },
         storageBucket: {
           authorization: true,
           documents: {
             authorization: true,
-            tagset: {
-              authorization: true,
-            },
+            tagset: { authorization: true },
           },
         },
       },
       select: {
         id: true,
+        authorization:
+          this.authorizationPolicyService.authorizationSelectOptions,
         tagsets: {
           id: true,
           authorization: