--- a/src/domain/community/organization/organization.service.authorization.ts
+++ b/src/domain/community/organization/organization.service.authorization.ts
@@ -98,7 +98,7 @@ export class OrganizationAuthorizationService {
     clonedOrganizationAuthorizationAnonymousAccess.anonymousReadAccess = true;
     const profileAuthorizations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        organization.profile,
+        organization.profile.id,
         clonedOrganizationAuthorizationAnonymousAccess
       );
     updatedAuthorizations.push(...profileAuthorizations);