--- a/src/domain/community/user-group/user-group.service.authorization.ts
+++ b/src/domain/community/user-group/user-group.service.authorization.ts
@@ -46,7 +46,7 @@ export class UserGroupAuthorizationService {
 
     const profileAuthoriations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        userGroup.profile,
+        userGroup.profile.id,
         userGroup.authorization
       );
     updatedAuthorizations.push(...profileAuthoriations);