--- a/src/domain/community/user/user.service.authorization.ts
+++ b/src/domain/community/user/user.service.authorization.ts
@@ -99,7 +99,7 @@ export class UserAuthorizationService {
     // cascade
     const profileAuthorizations =
       await this.profileAuthorizationService.applyAuthorizationPolicy(
-        user.profile,
+        user.profile.id,
         clonedAnonymousReadAccessAuthorization // Key that this is publicly visible
       );
     updatedAuthorizations.push(...profileAuthorizations);