… src/domain/community/virtual-contributor/virtual.contributor.service.authorization.ts
… src/domain/innovation-hub/innovation.hub.service.authorization.ts
… src/domain/space/space/space.service.authorization.ts
… src/domain/template/callout-template/callout.template.service.authorization.ts
… src/domain/template/community-guidelines-template/community.guidelines.template.service.authorization.ts
… src/domain/template/innovation-flow-template/innovation.flow.template.service.authorization.ts
… src/domain/template/post-template/post.template.service.authorization.ts
… src/domain/template/whiteboard-template/whiteboard.template.service.authorization.ts
… src/domain/timeline/event/event.service.authorization.ts
… src/library/innovation-pack/innovation.pack.service.authorization.ts
… src/platform/forum-discussion/discussion.service.authorization.ts