Optimizes the data that is loaded around Profiles.

Also done for CalloutFraming, CalloutContribution and User entity. 

Further optimization possible but this relatively small set of changes is showing some good results. On my local machine: 
- simple account with one space: auth reset went from 4.1s to 1.7s
- simple user: auth reset went from 380ms to 80ms

Suspect the win is across the board. More to be had, but given the above results thought it already interesting to push this.